requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.7.4
cfgv==3.4.0
distlib==0.3.8
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.15.4
h11==0.14.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.9.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.1
PyYAML==6.0.1
selenium==4.23.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
wsproto==1.2.0
